+++ NEWS.md
@@ -1,3 +1,8 @@
+# currently dev
+
+* Remove rgdal import. 
+
+
 # currently 0.1.0
 
 * Renamed from 'rk'.